+++ include/vyb/parser/parser.hpp
@@ -193,6 +193,9 @@ namespace vyb { // Changed Vyb to vyb
         std::unique_ptr<vyb::ast::IfStatement> parse_ensure();
         std::unique_ptr<vyb::ast::WhileStatement> parse_while();
         std::unique_ptr<vyb::ast::ForStatement> parse_for();
+        std::unique_ptr<vyb::ast::ForStatement> buildForLoopIteratorDesugar(
+            const SourceLocation& loc, const token::Token& ident, vyb::ast::ExprPtr range_expr,
+            std::unique_ptr<vyb::ast::BlockStatement> body, vyb::ast::ExprPtr skip_expr);
         std::unique_ptr<vyb::ast::ReturnStatement> parse_return();
         std::unique_ptr<vyb::ast::PassStatement> parse_pass();
         std::unique_ptr<vyb::ast::BreakStatement> parse_break();